rio: Reorganize RIO server code for DS operations
DS operations will start differing from MDS operations in
subtle and not so subtle ways, as we start dealing with
dirty inodes and such.

This commit segregates the notion of a common server library
and organizes the code to be able to start adding different
routines for the DS and MDS bricks as appropriate.

Change-Id: I4b808b9ecb387a4264ddc714d1e748160412d3b5
Signed-off-by: ShyamsundarR <srangana@redhat.com>