rio: Add iatt cleanup for iatt returned from DS
DS bricks are autoritative on size and m/atime information
only. They are not autoritative for any of the other fields
in the iatt structure.

Thus when returning iatt from the DS we need to filter the
rest of the information out, which is implemented in this
patch.

Also, in the future, if we could move to iatt2 or mask based
iatt values, to denote which fields are vaild (or vice-versa)
it would be a cleaner approach than this to provide the same
details back up the stack.

Change-Id: I700e1e0bd1732970cb36c0a2a2cb6ea457e8941d
Signed-off-by: ShyamsundarR <srangana@redhat.com>