rio: Added (s)setattr and (f)truncate support
(f)setattr was not handling dirty inodes, hence size
information returned was incorrect. This is fixed.

(f)truncate needs some special handling on the MDS when the
Op is just truncate, this is forwarded to the DS and then
returned back to the caller.

The changes are tested using glfsxmp program and also using
coreutils commands like truncate and stat.

Change-Id: I71328b5807e9808c2397c30b676714bac2b84a94
Signed-off-by: ShyamsundarR <srangana@redhat.com>